Fixed TimescaleDB Keys
TsDB requires that uniqe keys be used to shard the tables into
hypertables. This works fine when you want to use `timestamp` and a
couple other fields to enforce uniqueness. However this falls over when
you want to apply an ID field and use that as a reference on other
tables.

I removed the ID field from the datagrams table and the ref from the
measurements table. This may or may not fly, but for now it's how we get
this thing up an running.